refactoring and function inline

+++ tools/shell/include/linenoise.h
@@ -73,7 +73,7 @@ void linenoiseSetMultiLine(int ml);
 void linenoisePrintKeyCodes(void);
 void linenoiseMaskModeEnable(void);
 void linenoiseMaskModeDisable(void);
-size_t linenoiseComputeRenderWidth(const char* buf, size_t len);
+uint32_t linenoiseComputeRenderWidth(const char* buf, size_t len);
 
 #ifdef __cplusplus
 }
